[INTERPRETER] * BUG: Some 64 bits fixes.
[COMPILER]
* BUG: Fix how two error messages are printed.


git-svn-id: svn://localhost/gambas/trunk@1845 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/gbx/gbx_api.h
@@ -122,7 +122,7 @@ int GB_IsRightToLeft(void);
 
 void *GB_Eval(void *, void *);
 
-void GB_ArrayNew(GB_ARRAY *array, unsigned int type, int size);
+void GB_ArrayNew(GB_ARRAY *array, intptr_t type, int size);
 int GB_ArrayCount(GB_ARRAY array);
 void *GB_ArrayAdd(GB_ARRAY array);
 void *GB_ArrayGet(GB_ARRAY array, int index);
